♻️ Refactor TagList into TagListItem
